+++ app/sprinkles/account/config/debug.php
@@ -1,0 +1,12 @@
+<?php
+
+    /**
+     * Account configuration file for UserFrosting.
+     *
+     */
+
+    return [
+        'debug' => [
+            'auth' => true
+        ],
+    ];