+++ app/sprinkles/core/src/Bakery/SetupEnvCommand.php
@@ -82,6 +82,7 @@ class SetupEnvCommand extends BaseCommand
             $newEnvMode = $this->io->choice("Environment Mode", [
                 'default',
                 'production',
+                'debug',
                 'Other...'
             ], 'default');
         }